Migrate remaining tasks
