--- combine_task.ts
@@ -1,17 +1,0 @@
-const fs = require('fs');
-const path = require('path');
-const parse = require('csv-parse/lib/sync');
-const { reportToRow, reportToRowHeaders } = require('./report-to-row');
-const csvStringify = require('csv-stringify/lib/sync');
-const { aggregateCSVReports } = require('./combine');
-
-const reportsDirPath = process.argv[2];
-const outputDir = process.argv[3] || process.argv[2];
-const aggregatedReportData = aggregateCSVReports(reportsDirPath);
-const writePath = path.join(outputDir, 'aggregatedMobileReport.csv');
-fs.writeFile(writePath, aggregatedReportData, (e) => {
-  if (e) {
-    console.error(e);
-  }
-});
-console.log('DONE!');
