+++ package.json
@@ -6,8 +6,8 @@
     "test": "jest",
     "test:watch": "jest --watch",
     "urls": "node dist/urls-task.js",
-    "reports": "node dist/lighthouse_task.js",
-    "combine": "node dist/combine_task.js",
+    "reports": "node dist/lighthouse-task.js",
+    "combine": "node dist/combine-task.js",
     "scan": "node dist/scan-task.js",
     "check-lint": "eslint . && prettier --check .",
     "lint": "eslint --fix . && prettier --write .",
